Fixed type-conversion in eratosthenes_sieve

+++ algorithms/eratosthenes_sieve.h
@@ -4,9 +4,9 @@ std::vector<bool> build_eratosthenes_sieve(size_t n)
 {
     std::vector<bool> prime(n + 1, 1);
     prime[0] = prime[1] = 0;
-    for (int i = 2; i * i <= n; i++) {
+    for (size_t i = 2; i * i <= n; i++) {
         if (prime[i]) {
-            for (int j = 2; j <= n / i; j++) {
+            for (size_t j = 2; j <= n / i; j++) {
                 prime[i * j] = 0;
             }
         }
